flask_api/requirements.txt
# ── Core API ──────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# ── ML ────────────────────────────────────────────────────────────
scikit-learn==1.5.1
lightgbm==4.5.0
xgboost==2.1.1
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2

# ── MLflow ────────────────────────────────────────────────────────
mlflow==2.15.1
boto3==1.34.162          # S3 artifact store

# ── YouTube Data API ──────────────────────────────────────────────
google-api-python-client==2.139.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# ── Utils ─────────────────────────────────────────────────────────
pyyaml==6.0.2
python-dotenv==1.0.1
requests==2.32.3
